move br_set_state() to brmon.c
To unify all netlink related code in one unit, move br_set_state() to
brmon.c and unnexport rth_state.
